Code Changes
I decided to go a different direction with the code structure. The
original IAdult.cs was applied to the classes created by Entity
Framework. However, those classes are subject to be reset any time the
model changes, so I added an App_Code folder to my project and put these
new classes (IAdult.cs, Adult.cs, Child.cs, and
ContactInformationObject.cs) in it so that they won't be changed by a
code generating engine. Finally, I made IAdult.cs inherit IDisposable so
that we can have as few objects in memory as possible.