Added Files
I created Image files for the navigation buttons. I put all of the
various phase files for the image creation in the image folder to show
that I built the image files that will be used as navigation buttons.

I also included the SQL script that creates the Database that will
support this data intensive application.