Data rework
Reworking the database structures to create a third degree normalized
database. I also added a static class that will use ADO.NET to get data
for the drop down lists. Migrate.aspx is a data migration page that will
be used to pull in sample data and add any necessary fields to the
sample database.